# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Minha API",
        "description": "Documentação completa da minha API Laravel",
        "contact": {
            "name": "Suporte",
            "url": "https://example.com/suporte",
            "email": "suporte@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor local de desenvolvimento"
        },
        {
            "url": "https://api.example.com/api/v1",
            "description": "Servidor de produção"
        }
    ],
    "paths": {
        "/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Listar todas as reservas",
                "operationId": "573c89dfb714acb02c1b2f4826b22f0d",
                "responses": {
                    "200": {
                        "description": "Lista de reservas retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Reservation"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Criar uma nova reserva",
                "operationId": "dccb73c16710c16c1ee2391ae06b2e09",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "reservation_room_id",
                                    "first_name",
                                    "last_name",
                                    "hotel_id",
                                    "room_id",
                                    "arrival_date",
                                    "departure_date",
                                    "currencycode",
                                    "meal_plan",
                                    "totalprice",
                                    "guest_counts",
                                    "prices"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1001
                                    },
                                    "reservation_room_id": {
                                        "type": "integer",
                                        "example": 2001
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "João"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Silva"
                                    },
                                    "hotel_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "room_id": {
                                        "type": "integer",
                                        "example": 101
                                    },
                                    "arrival_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-10"
                                    },
                                    "departure_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-15"
                                    },
                                    "currencycode": {
                                        "type": "string",
                                        "example": "BRL"
                                    },
                                    "meal_plan": {
                                        "type": "string",
                                        "example": "breakfast"
                                    },
                                    "totalprice": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500
                                    },
                                    "guest_counts": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "type",
                                                "count"
                                            ],
                                            "properties": {
                                                "type": {
                                                    "type": "string",
                                                    "example": "adult"
                                                },
                                                "count": {
                                                    "type": "integer",
                                                    "minimum": 1,
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "prices": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "rate_id",
                                                "date",
                                                "amount"
                                            ],
                                            "properties": {
                                                "rate_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-06-10"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 300
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Quarto não disponível para o período"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/reservations/{reservation}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Exibir uma reserva específica",
                "operationId": "42f417a78cf996da1053c90767936fc4",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "ID da reserva",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1001
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva não encontrada"
                    }
                }
            }
        },
        "/reservations/{id}": {
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Remover uma reserva",
                "operationId": "7eb4cb182a79cf2e8697cf93a6bfc0f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da reserva",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1001
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva removida com sucesso"
                    },
                    "404": {
                        "description": "Reserva não encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Reservation": {
                "title": "Reservation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1001
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-06-01"
                    },
                    "time": {
                        "type": "string",
                        "example": "14:00:00"
                    },
                    "hotel_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "customer": {
                        "properties": {
                            "first_name": {
                                "type": "string",
                                "example": "João"
                            },
                            "last_name": {
                                "type": "string",
                                "example": "Silva"
                            }
                        },
                        "type": "object"
                    },
                    "room": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 101
                            },
                            "arrival_date": {
                                "type": "string",
                                "format": "date",
                                "example": "2024-06-10"
                            },
                            "departure_date": {
                                "type": "string",
                                "format": "date",
                                "example": "2024-06-15"
                            },
                            "currencycode": {
                                "type": "string",
                                "example": "BRL"
                            },
                            "meal_plan": {
                                "type": "string",
                                "example": "breakfast"
                            },
                            "totalprice": {
                                "type": "number",
                                "format": "float",
                                "example": 1500
                            },
                            "guest_counts": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "adult"
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "prices": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "rate_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-06-10"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 300
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}